create login view and template in member app

+++ django_app/post/views.py
@@ -36,14 +36,14 @@ def post_detail(request, post_pk):
     # ModelManager의 get 메서드를 통해 단 한 개의 객체만 가져온다.
     #   https://docs.djangoproject.com/en/1.11/ref/models/querysets/#get
 
-    # post = get_object_or_404(Post, pk=post_pk)
+    post = get_object_or_404(Post, pk=post_pk)
     # 가져오는 과정에서 예외처리하기
     #   - get()은 오브젝트가 없거나(DoesNotExist)
     #   - 여러 개의 오브젝트가 반환(MultipleObjectsReturns)될 때 에러가 난다
 
-    try:
-        post = Post.objects.get(pk=post_pk)
-    except Post.DoesNotExist as e:
+    # try:
+    #     post = Post.objects.get(pk=post_pk)
+    # except Post.DoesNotExist as e:
         # 1. 404에러를 띄워준다.
         # return HttpResponseNotFound('Post Not Found. detail: {}'.format(e))
 
@@ -56,33 +56,33 @@ def post_detail(request, post_pk):
         # 2-2. HttpResponseRedirect
         #   https://docs.djangoproject.com/en/1.11/ref/request-response/#django.http.HttpResponseRedirect
         # 템플릿에서 reverse함수로 url을 조합, 함수를 써주는 것과 위의 redirect()는 결국 같은 기능을 한다.
-        url = reverse('post:post_list')
-        return HttpResponseRedirect(url)
+        # url = reverse('post:post_list')
+        # return HttpResponseRedirect(url)
 
-    # request에 대해 response를 돌려줄 때는 HTTPResponse나 render함수 사용
+    # request에 대해 response를 돌려줄 때는 HttpResponse나 render함수 사용
     # template을 사용할 경우 render함수를 사용한다
     # render함수는...
     #   django.template.loader.get_template함수와
-    #   django.http.HTTPResponse함수를 축약해 놓은 shortcut
+    #   django.http.HttpResponse함수를 축약해 놓은 shortcut
 
     # ! 이 뷰에서는 render를 사용하지 않고 전체 과정(loader, HttpResponse)을 기술
     # Django가 템플릿을 검색할 수 있는 모든 디렉토리를 순회하며
     # 인자로 주어진 문자열 값과 일치하는 템플릿이 있는지 확인후
-    # 템플릿을 template에 리턴 (django.template.backends.django.Template클래스형 객체)
+    # 템플릿을 template에 리턴 (django.template.backends.django.Template 클래스형 객체)
     # get_template() 메서드
     #   https://docs.djangoproject.com/en/1.11/topics/templates/#django.template.loader.get_template
-    template = loader.get_template('post/post_detail.html')
+    # template = loader.get_template('post/post_detail.html')
     # dict형 변수 context의 'post'키에 post(Post객체) 할당
     context = {
         # context로 전달될 dictionary의 key값이 템플릿에서 사용가능한 변수명이 된다.
         'post': post,
     }
     # template에 인자로 주어진 context, request를 render함수로 string을 받아온다.
-    rendered_string = template.render(request=request, context=context)
-    return HTTPResponse(rendered_string)
+    # rendered_string = template.render(request=request, context=context)
+    # return HTTPResponse(rendered_string)
 
     # render를 사용하면 위의 과정을 함축하여 사용함과 동시에 post_detail.html으로 연결해준다.
-    # return render(request, 'post/post_detail.html', context=context)
+    return render(request, 'post/post_detail.html', context=context)
 
 
 def post_create(request):
@@ -103,7 +103,7 @@ def post_create(request):
         post = Post.objects.create(author=author, photo=photo)
 
         # POST 요청시 name이 'comment'인 input에서 전달된 값을 가져옴
-        # dic.get() : 값을
+        # dic.get() : 딕셔너리 내장함수로 값을 가져온다
         comment_string = request.POST.get('comment', '')
         # 빈 문자열 ''이나 None 모두 False로 평가되므로
         # if not 으로 댓글로 쓸 내용 또는 comment키가 전달되지 않았음을 검사 가능
@@ -132,19 +132,18 @@ def post_modify(request, post_pk):
         }
         return render(request, 'post/post_modify.html', context)
     else:
-        post.photo = request.FILES['file']
-        post.author = User.objects.first()
+        photo = request.FILES['file']
+        post.photo = photo
         post.save()
-        context = {
-            'post': post,
-        }
-        return render(request, 'post/post_detail.html', context)
+        return redirect('post:post_detail', post_pk=post.pk)
 
 
 def post_delete(request, post_pk):
-    post = Post.objects.get(pk=post_pk)
-    post.delete()
-    return redirect('post:post_list')
+    if request.method == "POST":
+        post = Post.objects.get(pk=post_pk)
+        post.delete()
+        return redirect('post:post_list')
+    return HTTPResponse('you have no authority to delete')
     # post_pk에 해당하는 Post에 대한 delete요청만을 받음
     # 처리완료 후에는 post_list페이지로 redirect
